requirements.txt
google-generativeai
SpeechRecognition
pyttsx3
python-dotenv
